+++ CHANGELOG.md
@@ -26,6 +26,8 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 * Added e3x.ops.normalize_and_return_norm
 * Added option to include mapping as weighting in mapped functions
 * Added option to return vector norms to e3x.nn.basis
+* Added e3x.nn.ExponentialBasis (wraps e3x.nn.basis, injecting learnable gamma)
+* Added e3x.ops.inverse_softplus helper function
 
 ## [1.0.1] - 2024-01-17